refactor: $store를 통해 todo, filter 접근 가능하도록 변경
 - 패키지 구조 변경, store/module
