+++ js/store/module/todo.js
@@ -1,6 +1,6 @@
-import storage from "./storage.js";
+import storage from "../storage.js";
 import filter from "./filter.js";
-import { FILTER } from "./constants.js";
+import { FILTER } from "../../constants.js";
 
 const todo = (() => {
   const items = storage.getStorage();
